+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add `.serializer` submatcher for relationship matchers (`belong_to`, `have_one` and `have_many`)
+
 ## [1.0.0] - 2021-04-10
 
 - Initial release